Remove support for embedding data in SkeletalRichEnum
Having fields is still possible in rich enums (including when
using SkeletalRichEnum).

This removal simplifies the current implementation and usage, while
also reducing footprint for the majority of rich enums, which rely on
SkeletalRichEnum and don't have embedded data.
The increased footprint can be avoided and is an artifact of the
way the removed functionality was implemented: an empty struct was used -
but even empty structs take up space, at least 1 byte, and then alignment
may increase the actual size further.

Downside: a type that uses SkeletalRichEnum as a base
while also defining additional fields in the child class
is no longer `std::is_standard_layout_v`. However, a scenario
where both embedding and standard_layout are requirements for an
optimal solution should be rare. If more such scenarios arise
in the future, this functionality can be restored (possibly
via a separate Skeletal base)